refactor: 컨트롤러 설명 수정
