+++ src/main/java/com/becareful/becarefulserver/domain/association/dto/JoinApplicationSimpleDto.java
@@ -1,8 +1,8 @@
 package com.becareful.becarefulserver.domain.association.dto;
 
 import com.becareful.becarefulserver.domain.association.domain.AssociationJoinApplication;
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
-import com.becareful.becarefulserver.domain.nursingInstitution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
