+++ src/main/java/com/becareful/becarefulserver/domain/association/service/AssociationService.java
@@ -15,7 +15,7 @@ import com.becareful.becarefulserver.domain.association.repository.AssociationRe
 import com.becareful.becarefulserver.domain.association.vo.AssociationJoinApplicationStatus;
 import com.becareful.becarefulserver.domain.community.domain.PostBoard;
 import com.becareful.becarefulserver.domain.community.repository.PostBoardRepository;
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
 import com.becareful.becarefulserver.domain.socialworker.repository.SocialWorkerRepository;
