+++ src/main/java/com/becareful/becarefulserver/domain/community/controller/CommentController.java
@@ -3,6 +3,8 @@ package com.becareful.becarefulserver.domain.community.controller;
 import com.becareful.becarefulserver.domain.community.dto.request.CommentCreateRequest;
 import com.becareful.becarefulserver.domain.community.dto.response.CommentResponse;
 import com.becareful.becarefulserver.domain.community.service.CommentService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -12,10 +14,12 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/community/board/{boardType}/post/{postId}/comment")
+@Tag(name = "Comment", description = "커뮤니티 탭 댓글 관련 API 입니다.")
 public class CommentController {
 
     private final CommentService commentService;
 
+    @Operation(summary = "댓글 작성")
     @PostMapping
     public ResponseEntity<Void> createComment(
             @PathVariable String boardType, @PathVariable Long postId, @RequestBody CommentCreateRequest request) {
@@ -25,6 +29,7 @@ public class CommentController {
                 .build();
     }
 
+    @Operation(summary = "댓글 조회")
     @GetMapping
     public ResponseEntity<List<CommentResponse>> getComments(
             @PathVariable String boardType, @PathVariable Long postId) {
