+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/response/InstitutionInfoResponse.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.matching.dto.response;
 
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
 
 public record InstitutionInfoResponse(String name, String address) {
     public static InstitutionInfoResponse from(NursingInstitution institution) {
