+++ src/main/java/com/becareful/becarefulserver/domain/nursing_institution/dto/response/NursingInstitutionCreateResponse.java
@@ -1,3 +1,3 @@
-package com.becareful.becarefulserver.domain.nursingInstitution.dto.response;
+package com.becareful.becarefulserver.domain.nursing_institution.dto.response;
 
 public record NursingInstitutionCreateResponse(Long institutionId, String institutionName) {}
