+++ src/main/java/com/becareful/becarefulserver/domain/nursing_institution/dto/response/NursingInstitutionSearchResponse.java
@@ -1,4 +1,4 @@
-package com.becareful.becarefulserver.domain.nursingInstitution.dto.response;
+package com.becareful.becarefulserver.domain.nursing_institution.dto.response;
 
 import java.util.List;
