--- src/main/java/com/becareful/becarefulserver/domain/socialworker/controller/NursingInstitutionController.java
@@ -1,1 +1,0 @@
-
