+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/repository/ElderlyRepository.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.socialworker.repository;
 
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.socialworker.domain.Elderly;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
