+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/repository/SocialWorkerRepository.java
@@ -1,7 +1,7 @@
 package com.becareful.becarefulserver.domain.socialworker.repository;
 
 import com.becareful.becarefulserver.domain.association.domain.Association;
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import java.util.List;
 import java.util.Optional;
