+++ src/test/java/com/becareful/becarefulserver/fixture/NursingInstitutionFixture.java
@@ -1,7 +1,7 @@
 package com.becareful.becarefulserver.fixture;
 
-import com.becareful.becarefulserver.domain.nursingInstitution.domain.NursingInstitution;
-import com.becareful.becarefulserver.domain.nursingInstitution.vo.FacilityType;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.NursingInstitution;
+import com.becareful.becarefulserver.domain.nursing_institution.vo.FacilityType;
 import java.util.EnumSet;
 
 public class NursingInstitutionFixture {
