+++ src/test/java/com/becareful/becarefulserver/fixture/SocialWorkerFixture.java
@@ -3,7 +3,7 @@ package com.becareful.becarefulserver.fixture;
 import static com.becareful.becarefulserver.fixture.NursingInstitutionFixture.NURSING_INSTITUTION;
 
 import com.becareful.becarefulserver.domain.common.vo.Gender;
-import com.becareful.becarefulserver.domain.nursingInstitution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
 import java.time.LocalDate;
